improve server perf and stability

+++ server/server.py
@@ -1,25 +1,28 @@
 import asyncio
 import os
 import sys
-import tempfile
 
-from fastapi import FastAPI, Request, HTTPException
+from fastapi import FastAPI, Request, HTTPException, BackgroundTasks
 from supabase import create_client, Client
 
 file_dir = os.path.dirname(__file__)
 sys.path.append(file_dir)
 
-from shared import ProcessVideoArgs
 from logger import log
+from extract_route import extract_route_multiprocessing
 import config
 
 supabase: Client = create_client(config.SUPABASE_URL, config.SUPABASE_SERVICE_KEY)
 
 app = FastAPI()
 
+routes_being_processed = set()
+
+process_video_lock = asyncio.Semaphore(1)
+
 
 @app.post('/process_video')
-async def process(request: Request):
+async def process(request: Request, background_tasks: BackgroundTasks):
     if request.headers.get("Authorization") != f'Bearer {config.SHARED_SECRET_KEY}':
         raise HTTPException(status_code=401, detail="Invalid auth token")
     log.info("Received request")
@@ -30,26 +33,37 @@ async def process(request: Request):
     if event["type"] != "INSERT" or event["record"]["bucket_id"] != "route_uploads":
         raise HTTPException(status_code=400, detail="Invalid event type")
 
-    asyncio.create_task(begin_processing(event))
+    async def begin_processing(event):
+        async with process_video_lock:
+            try:
+                routes_being_processed.add(event["record"]["id"])
+                await process_video(event)
+            except Exception as e:
+                route_id = os.path.basename(event["record"]["name"]).split(".")[0]
+                supabase.from_("routes").update({"status": "failed"}).eq("id", route_id).execute()
+                raise e
+            finally:
+                routes_being_processed.remove(event["record"]["id"])
+
+    background_tasks.add_task(begin_processing, event)
 
     return {"message": "Started processing"}, 202
 
 
-process_video_lock = asyncio.Semaphore(1)
+@app.post('/heatbeat')
+async def heartbeat(request: Request):
+    if request.headers.get("Authorization") != f'Bearer {config.SHARED_SECRET_KEY}':
+        raise HTTPException(status_code=401, detail="Invalid auth token")
+    log.info("Received heartbeat")
+    return {"message": "OK"}, 200
 
 
-async def begin_processing(event):
-    async with process_video_lock:
-        with tempfile.TemporaryDirectory(prefix="squad_gps") as tempdir:
-            await process_video(event, tempdir)
-
-
-async def process_video(data, tempdir):
-    _log = log.bind(remote_filepath=data["record"]["name"], tempdir=tempdir)
+async def process_video(data):
+    _log = log.bind(remote_filepath=data["record"]["name"])
     _log.info("Processing video")
 
     video_path = None
-    route_uploads_id = None
+    route_id = None
     try:
         # for some reason the download isn't completed by the time the webhook is called sometimes
         for i in range(20):
@@ -64,18 +78,14 @@ async def process_video(data, tempdir):
         else:
             raise HTTPException(status_code=500, detail="Failed to download video")
 
-        route_uploads_id = os.path.basename(data["record"]["name"]).split(".")[0]
-        upload_details = supabase.from_("route_uploads").select("*").eq("upload_id",
-                                                                        route_uploads_id).execute()
-        if not upload_details.data:
+        route_id = os.path.basename(data["record"]["name"]).split(".")[0]
+        route_details = supabase.from_("routes").select("*").eq("id", route_id).execute()
+
+        if not route_details.data:
             raise HTTPException(status_code=404, detail="Upload details not found")
 
-        upload_details = upload_details.data[0]
-        _log = _log.bind(upload_details=upload_details)
-
-        _log.info(f"getting route details for {upload_details['route_id']}")
-        route_details = supabase.from_("routes").select("*").eq("id",
-                                                                upload_details["route_id"]).single().execute().data
+        route_details = route_details.data[0]
+        _log = _log.bind(upload_details=route_details)
 
         video_path = os.path.join(config.DOWNLOADS_DIR, data["record"]["name"])
         _log.info(f"writing video to {video_path}")
@@ -84,50 +94,24 @@ async def process_video(data, tempdir):
         ## write to file
         with open(video_path, "wb") as f:
             f.write(bytes)
+        del bytes
 
-        script_path = os.path.join(os.path.dirname(__file__), "process_video.py")
-
-        tasks = []
-        outfiles = []
-        for i in range(config.NUM_PROCESSES):
-            filepath = os.path.join(tempdir, f"out_{i}.csv")
-            _log.info(f"starting process {i}")
-            sp_args = [script_path, *ProcessVideoArgs(map_name, video_path, str(i), filepath)]
-            tasks.append(asyncio.create_subprocess_shell(" ".join(sp_args)))
-            outfiles.append(filepath)
-
-        procs = await asyncio.gather(*tasks)
-
-        await asyncio.gather(*[proc.communicate() for proc in procs])
-
-        measurements = []
-        # parse outputs as csv
-        for outfile in outfiles:
-            _log.info(f"reading output from {outfile}")
-            with open(outfile, "r") as f:
-                output = f.read()
-                for line in output.strip().split("\n"):
-                    line = line.strip()
-                    if not line:
-                        continue
-                    x, y, time = line.split(',')
-                    measurements.append({"x": int(x), "y": int(y), "time": int(time)})
+        measurements = extract_route_multiprocessing(map_name, video_path, _log)
 
         _log.info(f"extracted {len(measurements)} measurements ")
         if len(measurements) > 0:
-            _log.info(f"Successfully processed {route_uploads_id} with status . Writing to db...")
-            supabase.from_("routes").update({"path": measurements}).eq("id", route_details["id"]).execute()
-            supabase.from_("route_uploads").update({"status": "success"}).eq("upload_id", route_uploads_id).execute()
+            _log.info(f"Successfully processed {route_id} with status . Writing to db...")
+            supabase.from_("routes").update({"path": measurements, "status": "success"}).eq("id", route_details[
+                "id"]).execute()
         else:
             _log.error("no measurements found")
-            supabase.from_("route_uploads").update({"status": "error"}).eq("upload_id", route_uploads_id).execute()
+            supabase.from_("routes").update({"status": "error"}).eq("id", route_id).execute()
 
     except Exception as e:
         log.exception(e)
-        if route_uploads_id:
-            supabase.from_("route_uploads").update({"status": "error"}).eq("upload_id",
-                                                                           route_uploads_id).execute()
+        if route_id:
+            supabase.from_("routes").update({"status": "error"}).eq("id", route_id).execute()
     finally:
-        if video_path is not None:
+        if video_path is not None and os.path.exists(video_path):
             os.remove(video_path)
         supabase.storage.from_("route_uploads").remove([data["record"]["name"]])
